+++ services/common/indexing/parsers.py
@@ -69,10 +69,15 @@ class Parser(ABC):
 
 class ParserInfoindexing(Parser):
     OPTIONAL_PARAMS = {
-        'windows_overlap': 100,
-        'windows_length': 300,
+        'window_overlap': 100,
+        'window_length': 300,
+        "windows": 1,
+        'sub_window_overlap': 50,
+        'sub_window_length': 100,
         'modify_index_docs': {}
     }
+
+    INDEXING_MODES = ["simple", "recursive", "surrounding_context_window"]
 
     MODEL_FORMAT = "infoindexing"
 
@@ -88,7 +93,6 @@ class ParserInfoindexing(Parser):
             self.logger.error("Error parsing JSON. No generic and specific configuration.",
                               exc_info=get_exc_info())
             raise ex
-
         try:
             self.get_dataset_csv_path()
         except KeyError as ex:
@@ -99,11 +103,6 @@ class ParserInfoindexing(Parser):
 
         try:
             self.get_project_conf()
-            self.get_url()
-            self.get_process_type()
-            self.get_process_id()
-            self.get_department()
-            self.get_csv()
         except KeyError as ex:
             self.logger.debug("Error parsing JSON. No configuration of project defined.",
                               exc_info=get_exc_info())
@@ -123,17 +122,13 @@ class ParserInfoindexing(Parser):
 
         try:
             self.get_index_conf()
-            self.get_index()
+            self.get_vector_storage_conf(vector_storages)
+            self.get_chunking_method()
             self.get_models(available_pools, available_models, models_credentials)
-
-            self.get_vector_storage(vector_storages)
-            self.get_windows_overlap()
-            self.get_windows_length()
-            self.get_modify_index_docs()
         except KeyError as ex:
             self.logger.debug(f"Error getting indexing params for {self.index_conf}",
                               exc_info=get_exc_info())
-            self.logger.error(f"[Process {self.dataset_status_key}] Error getting model params.",
+            self.logger.error(f"[Process {self.dataset_status_key}] Error getting indexation params.",
                               exc_info=get_exc_info())
             raise ex
 
@@ -153,22 +148,12 @@ class ParserInfoindexing(Parser):
         self.dataset_csv_path = self.generic.get('dataset_conf', {}).get('dataset_csv_path')
 
     def get_project_conf(self):
-        self.project_conf = get_project_config(generic=self.generic)
-
-    def get_url(self):
-        self.url = self.project_conf['report_url']
-
-    def get_process_type(self):
-        self.process_type = self.project_conf['process_type']
-
-    def get_process_id(self):
-        self.process_id = self.project_conf['process_id']
-
-    def get_department(self):
-        self.department = self.project_conf['department']
-
-    def get_csv(self):
-        self.csv = self.project_conf.get("csv", False)
+        project_conf = get_project_config(generic=self.generic)
+        self.url = project_conf['report_url']
+        self.process_type = project_conf['process_type']
+        self.process_id = project_conf['process_id']
+        self.department = project_conf['department']
+        self.csv = project_conf.get("csv", False)
 
     def get_do_titles(self):
         self.do_titles = get_do_titles(generic=self.generic)
@@ -179,8 +164,35 @@ class ParserInfoindexing(Parser):
     def get_index_conf(self):
         self.index_conf = self.generic['index_conf']
 
-    def get_index(self):
-        self.index = self.index_conf['index']
+    def get_vector_storage_conf(self, vector_storages):
+        vector_storage_conf = self.index_conf['vector_storage_conf']
+        self.index = vector_storage_conf['index']
+        self.modify_index_docs = vector_storage_conf.get('modify_index_docs', self.OPTIONAL_PARAMS['modify_index_docs'])
+        self.vector_storage = self.get_vector_storage(vector_storages, vector_storage_conf)
+
+    @staticmethod
+    def get_vector_storage(vector_storages, vector_storage_conf):
+        vector_storage = vector_storage_conf.get('vector_storage')
+        for vs in vector_storages:
+            if vs.get("vector_storage_name") == vector_storage:
+                return vs
+        raise PrintableGenaiError(400, f"Vector storage {vector_storage} not available")
+
+    def get_chunking_method(self):
+        self.chunking_method = self.index_conf['chunking_method']
+        if self.chunking_method.get('method') in self.INDEXING_MODES:
+            # If exists get the value if not, set the default value
+            self.chunking_method.setdefault('window_overlap', self.OPTIONAL_PARAMS['window_overlap'])
+            self.chunking_method.setdefault('window_length', self.OPTIONAL_PARAMS['window_length'])
+            if self.chunking_method['method'] == "recursive":
+                self.chunking_method.setdefault('sub_window_overlap', self.OPTIONAL_PARAMS['sub_window_overlap'])
+                self.chunking_method.setdefault('sub_window_length', self.OPTIONAL_PARAMS['sub_window_length'])
+            elif self.chunking_method['method'] == "surrounding_context_window":
+                self.chunking_method.setdefault('windows', self.OPTIONAL_PARAMS['windows'])
+        else:
+            raise PrintableGenaiError(400, f"Chunking mode {self.chunking_method.get('method')} not available")
+
+
 
     def get_models(self, available_pools: dict, available_models: dict, models_credentials: dict):
         self.models = self.index_conf['models']
@@ -206,23 +218,6 @@ class ParserInfoindexing(Parser):
             if model.get('embedding_model') in unique_embedding_models:
                 raise PrintableGenaiError(400, f"Model '{model.get('embedding_model')}' duplicated")
             unique_embedding_models.append(model.get('embedding_model'))
-
-    def get_vector_storage(self, vector_storages):
-        vector_storage = self.index_conf.get('vector_storage')
-        for vs in vector_storages:
-            if vs.get("vector_storage_name") == vector_storage:
-                self.vector_storage = vs
-                return
-        raise PrintableGenaiError(400, f"Vector storage {vector_storage} not available")
-
-    def get_windows_overlap(self):
-        self.windows_overlap = self.index_conf.get('windows_overlap', self.OPTIONAL_PARAMS['windows_overlap'])
-
-    def get_windows_length(self):
-        self.windows_length = self.index_conf.get('windows_length', self.OPTIONAL_PARAMS['windows_length'])
-
-    def get_modify_index_docs(self):
-        self.modify_index_docs = self.index_conf.get('modify_index_docs', self.OPTIONAL_PARAMS['modify_index_docs'])
 
 
 class ParserInforetrieval(Parser):